requirements.txt
django
djangorestframework
dj-rest-auth

